refactor: replaced strings file with constants
