--- packages/dev-frontend/src/strings.ts
@@ -1,2 +1,0 @@
-export const COIN = "thUSD";
-export const FIRST_ERC20_COLLATERAL = "WETH";
